fix: force int for array indices #19

+++ MixtComp/src/lib/Mixture/Functional/FunctionCS.cpp
@@ -9,7 +9,7 @@
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Affero General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>
  **/
@@ -242,7 +242,7 @@ void FunctionCS::quantile(Vector<Real>& quantile) {
 	quantile(nQuantile - 1) = sortedT(nTime_ - 1);
 
 	for (Index q = 1; q < nQuantile - 1; ++q) {
-		quantile(q) = sortedT(q * quantileSize * (nTime_ - 1));
+		quantile(q) = sortedT(int(q * quantileSize * (nTime_ - 1)));
 	}
 }
